% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_custom_intervals}
\alias{calculate_custom_intervals}
\title{Calculate custom intervals when map_color_scale = "Custom"}
\usage{
calculate_custom_intervals(cutoff_points, domain)
}
\description{
Calculate custom intervals when map_color_scale = "Custom"
}
