% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_legend_cuts}
\alias{create_legend_cuts}
\title{Create cutoff points for legends for bubbles maps}
\usage{
create_legend_cuts(x, number = seq(1:10))
}
\description{
Create cutoff points for legends for bubbles maps
}
