% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{lflt_bubbles_CatGlnGlt}
\alias{lflt_bubbles_CatGlnGlt}
\title{Leaflet bubbles by category}
\usage{
lflt_bubbles_CatGlnGlt(
  data = NULL,
  mapName = "world_countries",
  opts = NULL,
  ...
)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbles by category
}
\section{ctypes}{
 Gcd-Num
}

\examples{
lflt_bubbles_CatGlnGlt(sampleData("Cat-Gln-Glt", nrow = 10))
}
