% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{lflt_bubbles_CatGlnGltNum}
\alias{lflt_bubbles_CatGlnGltNum}
\title{Leaflet bubbles by category size}
\usage{
lflt_bubbles_CatGlnGltNum(data = NULL, mapName = "world_countries",
  opts = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbles by category size
}
\section{ctypes}{
 Cat-Gln-Glt-Num
}

\examples{
lflt_bubbles_CatGlnGltNum(sampleData("Cat-Gln-Glt-Num", nrow = 10))
}
