% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles_GcdCatNum.R
\name{lflt_bubbles_GcdCat}
\alias{lflt_bubbles_GcdCat}
\title{Leaflet bubbles by categorical variable}
\usage{
lflt_bubbles_GcdCat(data = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbles by categorical variable
}
\section{ctypes}{
 Gcd-Cat
}

\examples{
lflt_bubbles_GcdCat(sample_data("Gcd-Cat", nrow = 10))
}
