% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles_GcdCatNum.R
\name{lflt_bubbles_GcdCatNum}
\alias{lflt_bubbles_GcdCatNum}
\title{Bubbles chart Gcd Cat Num}
\usage{
lflt_bubbles_GcdCatNum(data = NULL, ...)
}
\description{
\code{lflt_bubbles_GcdCatNum()} Create a Leaflet bubbles map based on a particular data type.
In this case, you can load data with only three columns, where the firts it's a \strong{geocode column},
the second is a \strong{categorical column} and the third must be  a \strong{numeric class column}, you should be sure that
three firts columns they meet this condition
}
\section{Ftype}{

Gcd-Cat-Num
}

\examples{
data <- sample_data("Gcd-Cat-Num", n = 30)
lflt_bubbles_GcdCatNum(data)

# Activate data labels
lflt_bubbles_GcdCatNum(data,
                       dataLabels_show = TRUE)

# if you want to calculate the average instead of the sum, you can use agg inside a function
lflt_bubbles_GcdCatNum(data,
                       agg = "mean",
                       dataLabels_show = TRUE)

# data with more of one column
data <- sample_data("Gcd-Cat-Num-Dat-Yea-Cat", n = 30)
lflt_bubbles_GcdCatNum(data)

# Change variable to color and pallete type
lflt_bubbles_GcdCatNum(data,
                       color_by = names(data)[2],
                       palette_type = "sequential")

# Change tooltip info and add additional information contained in your data
names_data <- names(data)
info_tool <- paste0("<b>",names_data[1],":</b> {", names_data[1],"}<br/><b>", names_data[4],":</b> {", names_data[4],"}<br/>")
data \%>\%
 lflt_bubbles_GcdCatNum(tooltip = info_tool)
}
\seealso{
Other Gcd-Cat-Num plots: 
\code{\link{lflt_choropleth_GcdCatNum}()}
}
\concept{Gcd-Cat-Num plots}
