% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles_GcdNum.R
\name{lflt_bubbles_GcdNum}
\alias{lflt_bubbles_GcdNum}
\title{Leaflet bubbless by numerical variable}
\usage{
lflt_bubbles_GcdNum(data = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbless by numerical variable
}
\section{ctypes}{
 Gcd-Num
}

\examples{
lflt_bubbles_GcdNum(sampleData("Gcd-Num", nrow = 10))
}
