% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{lflt_bubbles_GeNu}
\alias{lflt_bubbles_GeNu}
\title{lflt_bubbles_GeNu
Multilines}
\usage{
lflt_bubbles_GeNu(data, geoinfoPath = NULL, geoCodeVar = NULL,
  geoNameVar = NULL, bounds = NULL)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
lflt_bubbles_GeNu
Multilines
}
\section{ftypes}{
 Ye-Nu*
}

\examples{
lflt_bubbles_GeNu(sampleData("Ge-Nu",nrow = 10))
}
