% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{lflt_bubbles_GeNuNu}
\alias{lflt_bubbles_GeNuNu}
\title{lflt_bubbles_GeNuNu
Multilines}
\usage{
lflt_bubbles_GeNuNu(data, geoinfoPath = NULL, geoCodeVar = NULL,
  geoNameVar = NULL, bounds = NULL, pos_leg = "bottomright")
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
lflt_bubbles_GeNuNu
Multilines
}
\section{ftypes}{
 Ye-Nu*
}
\examples{
lflt_bubbles_GeNuNu(sampleData("Ge-Nu-Nu",nrow = 10))
}

