% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles_GlnGltCatNum.R
\name{lflt_bubbles_GlnGltCat}
\alias{lflt_bubbles_GlnGltCat}
\title{Leaflet bubbles by categorical variable}
\usage{
lflt_bubbles_GlnGltCat(data = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbles by categorical variable
}
\section{ctypes}{
 Gln-Glt-Cat
}

\examples{
lflt_bubbles_GlnGltCat(sample_data("Gln-Glt-Cat", nrow = 10))
}
