% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles_GlnGltCatNum.R
\name{lflt_bubbles_GlnGltCatNum}
\alias{lflt_bubbles_GlnGltCatNum}
\title{Leaflet bubbles by categorical variable}
\usage{
lflt_bubbles_GlnGltCatNum(data = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbles by categorical variable
}
\section{ctypes}{
 Gln-Glt-Cat-Num
}

\examples{
lflt_bubbles_GlnGltCatNum(sample_data("Gln-Glt-Cat-Num", nrow = 10))
}
