% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{lflt_bubbles_GlnGltNum}
\alias{lflt_bubbles_GlnGltNum}
\title{Leaflet bubbles}
\usage{
lflt_bubbles_GlnGltNum(data = NULL, mapName = "world_countries",
  opts = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbles
}
\section{ctypes}{
 Gln-Glt-Num
}

\examples{
lflt_bubbles_GlnGltNum(sampleData("Gln-Glt-Num", nrow = 10))
}
