% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles_GnmNum.R
\name{lflt_bubbles_Gnm}
\alias{lflt_bubbles_Gnm}
\title{Leaflet bubbless by numerical variable}
\usage{
lflt_bubbles_Gnm(data = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbless by numerical variable
}
\section{ctypes}{
 Gnm
}

\examples{
lflt_bubbles_Gnm(sampleData("Gnm", nrow = 10))
}
