% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles_GnmCatNum.R
\name{lflt_bubbles_GnmCat}
\alias{lflt_bubbles_GnmCat}
\title{Leaflet bubbles by categorical variable}
\usage{
lflt_bubbles_GnmCat(data = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbles by categorical variable
}
\section{ctypes}{
 Gnm-Cat
}

\examples{
lflt_bubbles_GnmCat(sample_data("Gnm-Cat", nrow = 10))
}
