% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles_GnmNum.R
\name{lflt_bubbles_GnmNum}
\alias{lflt_bubbles_GnmNum}
\title{Bubbles chart Gnm Num}
\usage{
lflt_bubbles_GnmNum(data = NULL, ...)
}
\description{
\code{lflt_bubbles_GnmNum()} Create a Leaflet bubbles map based on a particular data type.
In this case, you can load data with only two columns, where the firts it's a \strong{geoname column},
and the second is a \strong{numeric class column}, or make sure that the first two columns of
your data meet this condition
}
\section{Ftype}{

Gnm-Num
}

\examples{
data <- sample_data("Gnm-Num", n = 30)
lflt_bubbles_GnmNum(data)


# if you want to calculate the average instead of the sum, you can use agg inside a function
lflt_bubbles_GnmNum(data,
                       agg = "mean")


}
\seealso{
Other Gnm-Num plots: 
\code{\link{lflt_choropleth_GnmNum}()}
}
\concept{Gnm-Num plots}
