% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{lflt_bubbles_grouped_GcdCat}
\alias{lflt_bubbles_grouped_GcdCat}
\title{Leaflet bubbles grouped by categorical variable}
\usage{
lflt_bubbles_grouped_GcdCat(data, palette = c("#009EE3", "#9B71AF",
  "#FFFF99"), fillOpacity = 0.5, label = NULL, popup = NULL,
  size = 5, scope = "world_countries", tiles = "CartoDB.Positron")
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbles grouped by categorical variable
}
\section{ctypes}{
 Gcd-Cat
}

\examples{
lflt_bubbles_grouped_GcdCat(sampleData("Gcd-Cat", nrow = 10))
}
