% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{lflt_bubbles_grouped_GlnGltCat}
\alias{lflt_bubbles_grouped_GlnGltCat}
\title{Leaflet bubbles colored by categorical variable}
\usage{
lflt_bubbles_grouped_GlnGltCat(data, palette = c("#009EE3", "#9B71AF",
  "#FFFF99"), fillOpacity = 0.5, label = NULL, popup = NULL,
  size = 5, tiles = "CartoDB.Positron")
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbles colored by categorical variable
}
\section{ctypes}{
 Gln-Glt-Cat
}

\examples{
lflt_bubbles_grouped_GlnGltCat(sampleData("Gln-Glt-Cat", nrow = 10))
}
