% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{lflt_bubbles_grouped_GnmCat}
\alias{lflt_bubbles_grouped_GnmCat}
\title{Leaflet bubbles grouped by categorical variable}
\usage{
lflt_bubbles_grouped_GnmCat(data, palette = c("#009EE3", "#9B71AF",
  "#FFFF99"), fillOpacity = 0.5, label = NULL, popup = NULL, size = 5,
  scope = "world_countries", tiles = "CartoDB.Positron")
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbles grouped by categorical variable
}
\section{ctypes}{
 Gnm-Cat
}

\examples{
lflt_bubbles_grouped_GnmCat(sampleData("Gnm-Cat", nrow = 10))
}
