% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{lflt_bubbles_size_Gcd}
\alias{lflt_bubbles_size_Gcd}
\title{Leaflet bubbles size by geographical code}
\usage{
lflt_bubbles_size_Gcd(data, color = "navy", fillOpacity = 0.5,
  label = NULL, popup = NULL, minSize = 3, maxSize = 20,
  scope = "world_countries", tiles = "CartoDB.Positron")
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbles size by geographical code
}
\section{ctypes}{
 Gcd
}

\examples{
lflt_bubbles_size_Gcd(sampleData("Gcd", nrow = 10))
}
