% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{lflt_bubbles_size_GcdNum}
\alias{lflt_bubbles_size_GcdNum}
\alias{lflt_bubbles_GcdNum}
\title{Leaflet bubbles size by numeric variable}
\usage{
lflt_bubbles_GcdNum(data = NULL, mapName = "world_countries",
  opts = NULL)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbles size by numeric variable
}
\section{ctypes}{
 Gcd-Num
}

\examples{
lflt_bubbles_size_GcdNum(sampleData("Gcd-Num", nrow = 10))
}
