% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{lflt_bubbles_size_GlnGltCat}
\alias{lflt_bubbles_size_GlnGltCat}
\title{Leaflet bubbles size by categorical variable}
\usage{
lflt_bubbles_size_GlnGltCat(data, color = "navy", fillOpacity = 0.5,
  label = NULL, popup = NULL, minSize = 3, maxSize = 20,
  tiles = "CartoDB.Positron")
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbles size by categorical variable
}
\section{ctypes}{
 Gln-Glt-Cat
}

\examples{
lflt_bubbles_size_GlnGltCat(sampleData("Gln-Glt-Cat", nrow = 10))
}
