% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{lflt_bubbles_size_GltLn}
\alias{lflt_bubbles_size_GltLn}
\title{lflt_bubbles_size_GltLn
Multilines}
\usage{
lflt_bubbles_size_GltLn(data, bounds = NULL)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
lflt_bubbles_size_GltLn
Multilines
}
\section{ftypes}{
 Ye-Nu*
}

\examples{
lflt_bubbles_size_GltLn(sampleData("Gcd-Num",nrow = 10))
}
