% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{lflt_bubbles_size_Gnm}
\alias{lflt_bubbles_size_Gnm}
\title{Leaflet bubbles size by geographical name}
\usage{
lflt_bubbles_size_Gnm(data, color = "navy", fillOpacity = 0.5,
  label = NULL, popup = NULL, minSize = 3, maxSize = 20,
  scope = "world_countries", tiles = "CartoDB.Positron")
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbles size by geographical name
}
\section{ctypes}{
 Gnm
}

\examples{
lflt_bubbles_size_Gnm(sampleData("Gnm", nrow = 10))
}
