% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleths.R
\name{lflt_choropleth_Gcd}
\alias{lflt_choropleth_Gcd}
\title{Leaflet choropleths by geographical code}
\usage{
lflt_choropleth_Gcd(data = NULL, caption = NULL, count = TRUE,
  border = list(), fill = list(), format = c("", ""),
  labelWrap = 12, legend = list(), mapName = "world_countries",
  marks = c(",", "."), nDigits = 2, label = NULL,
  percentage = FALSE, popup = NULL, tiles = NULL)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet choropleths by geographical code
}
\section{ctypes}{
 Gcd
}

\examples{
lflt_choropleth_Gcd(sampleData("Gcd", nrow = 10))
}
