% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\name{lflt_choropleth_GcdCat}
\alias{lflt_choropleth_GcdCat}
\title{Leaflet choropleths by geographical variable}
\usage{
lflt_choropleth_GcdCat(data = NULL, mapName = "world_countries",
  opts = NULL)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet choropleths by geographical variable
}
\section{ctypes}{
 Gcd-Cat
}

\examples{
lflt_choropleth_GcdCat(sampleData("Gcd-Cat", nrow = 10))
}
