% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleths.R
\name{lflt_choropleth_Gnm}
\alias{lflt_choropleth_Gnm}
\title{Leaflet choropleths by geographical name}
\usage{
lflt_choropleth_Gnm(data, fillOpacity = 0.5, format = c("", ""),
  legendColor = "discrete", legendPosition = "bottomleft",
  legendTitle = NULL, label = NULL, marks = c(",", "."),
  matchMethod = "lv", maxDist = 2, nBins = 4, nDigits = 2,
  palette = c("#009EE3", "#E5007D", "#95C11E"), percentage = FALSE,
  scope = "world_countries", tiles = "CartoDB.Positron")
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet choropleths by geographical name
}
\section{ctypes}{
 Gnm
}

\examples{
lflt_choropleth_Gnm(sampleData("Gnm", nrow = 10))
}
