% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\name{lflt_choropleth_GnmCat}
\alias{lflt_choropleth_GnmCat}
\title{Leaflet choropleths by geographical variable}
\usage{
lflt_choropleth_GnmCat(data = NULL, mapName = "world_countries",
  opts = NULL)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet choropleths by geographical variable
}
\section{ctypes}{
 Gnm-Cat
}

\examples{
lflt_choropleth_GnmCat(sampleData("Gnm-Cat", nrow = 10))
}
