% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{lflt_heatmap_Gcd}
\alias{lflt_heatmap_Gcd}
\title{Leaflet heatmap by geographical code}
\usage{
lflt_heatmap_Gcd(data, radius = NULL, blur = NULL, gradient = NULL,
  tiles = "CartoDB.Positron", scope = "world_countries")
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet heatmap by geographical code
}
\section{ctypes}{
 Gcd
}

\examples{
lflt_heatmap_Gcd(sampleData("Gcd", nrow = 10))
}
