% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_GlnGltNum.R
\name{lflt_heatmap_GlnGlt}
\alias{lflt_heatmap_GlnGlt}
\title{#' Heatmap chart Gln Glt}
\usage{
lflt_heatmap_GlnGlt(data = NULL, ...)
}
\description{
\code{lflt_heatmap_GlnGlt()} Create a Leaflet bubbles map based on a particular data type.
In this case, you can load data with only two columns, where the firts column contains \strong{longitudes points},
and the second \strong{latitudes points}, or make sure that the first two columns of
your data meet this condition
}
\section{Ftype}{

Gln-Glt
}

\examples{
data <- sample_data("Gln-Glt", n = 30)
lflt_heatmap_GlnGlt(data)
}
\seealso{
Other Gln-Glt plots: 
\code{\link{lflt_hexmap_GlnGlt}()}
}
\concept{Gln-Glt plots}
