% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_GlnGltNum.R
\name{lflt_heatmap_GlnGltNum}
\alias{lflt_heatmap_GlnGltNum}
\title{#' Heatmap chart Gln Glt Num}
\usage{
lflt_heatmap_GlnGltNum(data = NULL, ...)

lflt_heatmap_GlnGltNum(data = NULL, ...)
}
\description{
\code{lflt_heatmap_GlnGltNum()} Create a Leaflet Heatmap map based on a particular data type.
In this case, you can load data with only three columns, where the firts column contains \strong{longitudes points},
the second \strong{latitudes points} and the third must be a \strong{numeric class column},
or be sure that three firts columns they meet this condition.

\code{lflt_heatmap_GlnGlt()} Create a Leaflet bubbles map based on a particular data type.
In this case, you can load data with only two columns, where the firts column contains \strong{longitudes points},
and the second \strong{latitudes points}, or make sure that the first two columns of
your data meet this condition
}
\section{Ftype}{

Gln-Glt-Num


Gln-Glt
}

\examples{
data <- sample_data("Gln-Glt-Num", n = 30)
lflt_Heatmap_GlnGltNum(data)
data <- sample_data("Gln-Glt", n = 30)
lflt_Heatmap_GlnGlt(data)
}
\seealso{
Other Gln-Glt-Num plots: 
\code{\link{lflt_bubbles_GlnGltNum}()}
}
\concept{Gln-Glt plots}
\concept{Gln-Glt-Num plots}
