% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{lflt_heatmap_GltGln}
\alias{lflt_heatmap_GltGln}
\title{lflt_heatmap_GltGln
Heatmap}
\usage{
lflt_heatmap_GltGln(data, radius = NULL)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
lflt_heatmap_GltGln
Heatmap
}
\section{ftypes}{
 Glt-Gln
}

\examples{
lflt_heatmap_size_GltGln(sampleData("Glt-Gln",nrow = 10))
}
