% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{lflt_heatmap_Gnm}
\alias{lflt_heatmap_Gnm}
\title{Leaflet heatmap by geographical name}
\usage{
lflt_heatmap_Gnm(data, radius = NULL, blur = NULL, gradient = NULL,
  tiles = "CartoDB.Positron", scope = "world_countries")
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet heatmap by geographical name
}
\details{
lflt_heatmap_Gnm
Heatmap
}
\section{ctypes}{
 Gnm
}

\examples{
lflt_heatmap_Gnm(sampleData("Gnm", nrow = 10))
}
