% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{lflt_markers_GlnGlt}
\alias{lflt_markers_GlnGlt}
\title{Leaflet markers by latitud and longitud}
\usage{
lflt_markers_GlnGlt(data, label = NULL, popup = NULL, icon = NULL,
  iconWidth = 36, iconHeight = 24, tiles = "CartoDB.Positron")
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet markers by latitud and longitud
}
\section{ctypes}{
 Gln-Glt
}

\examples{
lflt_markers_GlnGlt(sampleData("Gln-Glt", nrow = 10))
}
