% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{lflt_markers_GlnGltCat}
\alias{lflt_markers_GlnGltCat}
\title{Leaflet markers by latitud and longitud and icons by categorical variable}
\usage{
lflt_markers_GlnGltCat(data, label = NULL, popup = NULL, icon = NULL,
  iconWidth = 36, iconHeight = 24, tiles = "CartoDB.Positron")
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet markers by latitud and longitud and icons by categorical variable
}
\section{ctypes}{
 Gln-Glt-Cat
}

\examples{
lflt_markers_GlnGltCat(sampleData("Gln-Glt-Cat", nrow = 10))
}
