% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{lflt_markers_GlnGltImg}
\alias{lflt_markers_GlnGltImg}
\title{Leaflet markers by latitud and longitud and icons}
\usage{
lflt_markers_GlnGltImg(data, label = NULL, popup = NULL, iconWidth = 36,
  iconHeight = 24, tiles = "CartoDB.Positron")
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet markers by latitud and longitud and icons
}
\section{ctypes}{
 Gln-Glt-Img
}

\examples{
lflt_markers_GlnGltImg(sampleData("Gln-Glt-Img", nrow = 10))
}
