% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{lflt_bubbles_Gcd}
\alias{lflt_bubbles_Gcd}
\title{Leaflet bubbles geo code}
\usage{
lflt_bubbles_Gcd(data = NULL, mapName = "world_countries",
  opts = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbles
}
\section{ctypes}{
 Gcd
}

\examples{
lflt_bubbles_Gcd(sampleData("Gcd", nrow = 10))
}
