% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles_GcdNum.R
\name{lflt_bubbles_Gcd}
\alias{lflt_bubbles_Gcd}
\title{Leaflet bubbless by numerical variable}
\usage{
lflt_bubbles_Gcd(data = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbless by numerical variable
}
\section{ctypes}{
 Gcd
}

\examples{
lflt_bubbles_Gcd(sampleData("Gcd", nrow = 10))
}
