% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles_GcdCatNum.R
\name{lflt_bubbles_GcdCatNum}
\alias{lflt_bubbles_GcdCatNum}
\title{Leaflet bubbles by categorical variable}
\usage{
lflt_bubbles_GcdCatNum(data = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbles by categorical variable
}
\section{ctypes}{
 Gcd-Cat-Num
}

\examples{
lflt_bubbles_GcdCatNum(sample_data("Gcd-Cat-Num", nrow = 10))
}
