% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles_GlnGlt.R
\name{lflt_bubbles_GlnGlt}
\alias{lflt_bubbles_GlnGlt}
\title{Leaflet bubbless by numerical variable}
\usage{
lflt_bubbles_GlnGlt(data = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbless by numerical variable
}
\section{ctypes}{
 Gnm-Num
}

\examples{
lflt_bubbles_GlnGlt(sampleData("Gnm-Num", nrow = 10))
}
