% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{lflt_bubbles_GlnGlt}
\alias{lflt_bubbles_GlnGlt}
\title{Leaflet bubbles by latitud and longitud}
\usage{
lflt_bubbles_GlnGlt(data, caption = NULL, count = TRUE,
  border = list(weight = 1.3, color = "black", opacity = 1, stroke =
  TRUE), fill = list(color = NULL, opacity = 0.5, scale = "discrete",
  nullColor = "#dddddd"), format = c("", ""), highlightValue = NULL,
  highlightValueColor = NULL, labelWrap = 12, legend = list(bins = 6,
  position = "bottomleft", title = NULL), marks = c(",", "."),
  nDigits = 2, label = NULL, percentage = FALSE, popup = NULL,
  size = c(3, 20), tiles = NULL)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbles by latitud and longitud
}
\section{ctypes}{
 Gln-Glt
}

\examples{
lflt_bubbles_GlnGlt(sampleData("Gln-Glt", nrow = 10))
}
