% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{lflt_bubbles_GlnGlt}
\alias{lflt_bubbles_GlnGlt}
\title{Leaflet bubbles}
\usage{
lflt_bubbles_GlnGlt(data = NULL, mapName = "world_countries",
  opts = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbles
}
\section{ctypes}{
 Gln-Glt
}

\examples{
lflt_bubbles_GlnGlt(sampleData("Gln-Glt", nrow = 10))
}
