% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{lflt_bubbles_GlnGlt}
\alias{lflt_bubbles_GlnGlt}
\title{Leaflet bubbles by latitud and longitud}
\usage{
lflt_bubbles_GlnGlt(data, caption = NULL, count = TRUE,
  border = list(), fill = list(), format = c("", ""),
  labelWrap = 12, legend = list(), marks = c(",", "."),
  nDigits = 2, label = NULL, percentage = FALSE, popup = NULL,
  size = c(3, 20), tiles = NULL)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbles by latitud and longitud
}
\section{ctypes}{
 Gln-Glt
}

\examples{
lflt_bubbles_GlnGlt(sampleData("Gln-Glt", nrow = 10))
}
