% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{lflt_bubbles_GlnGltCat}
\alias{lflt_bubbles_GlnGltCat}
\title{Leaflet bubbles by latitud and longitud and category}
\usage{
lflt_bubbles_GlnGltCat(data, caption = NULL, count = TRUE,
  border = list(), fill = list(), format = c("", ""),
  labelWrap = 12, marks = c(",", "."), nDigits = 2, label = NULL,
  percentage = FALSE, popup = NULL, size = c(3, 20), tiles = NULL)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbles by latitud and longitud and category
}
\section{ctypes}{
 Gln-Glt-Cat
}

\examples{
lflt_bubbles_GlnGltCat(sampleData("Gln-Glt-Cat", nrow = 10))
}
