% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{lflt_bubbles_GlnGltCatNum}
\alias{lflt_bubbles_GlnGltCatNum}
\title{Leaflet bubbles grouped by categorical variable size by numerical}
\usage{
lflt_bubbles_GlnGltCatNum(data, caption = NULL, agg = "sum",
  border = list(), fill = list(), format = c("", ""),
  labelWrap = 12, marks = c(",", "."), nDigits = 2, label = NULL,
  percentage = FALSE, popup = NULL, size = c(3, 20), tiles = NULL)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbles grouped by categorical variable size by numerical
}
\section{ctypes}{
 Gln-Glt-Cat-Num
}

\examples{
lflt_bubbles_GlnGltCatNum(sampleData("Gln-Glt-Cat-Num", nrow = 10))
}
