% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{lflt_bubbles_GlnGltNum}
\alias{lflt_bubbles_GlnGltNum}
\title{Leaflet bubbles}
\usage{
lflt_bubbles_GlnGltNum(
  data,
  map_name = "world_countries",
  add_tiles = TRUE,
  tile_name = NULL,
  topo_weight = 1,
  topo_color = "#CCCCCC",
  topo_fill = FALSE,
  legend_show = TRUE,
  popup = NULL,
  colors = c("#f0f05a", "#f03f4e"),
  opts = NULL,
  ...
)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbles
}
\section{ctypes}{
 Gln-Glt-Num
}

\examples{
lflt_bubbles_GlnGltNum(sampleData("Gln-Glt-Num", nrow = 10))
}
