% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles_GlnGlt.R
\name{lflt_bubbles_GlnGltNum}
\alias{lflt_bubbles_GlnGltNum}
\title{Leaflet bubbles by numerical variable}
\usage{
lflt_bubbles_GlnGltNum(data = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbles by numerical variable
}
\section{ctypes}{
 Gnm-Num
}

\examples{
lflt_bubbles_GlnGltNum(sampleData("Gln-Glt-Num", nrow = 10))
}
