% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{lflt_bubbles_GlnGltNum}
\alias{lflt_bubbles_GlnGltNum}
\title{Leaflet bubbles by latitud and longitud and numeric variable}
\usage{
lflt_bubbles_GlnGltNum(data, caption = NULL, agg = "sum",
  border = list(), fill = list(), format = c("", ""),
  labelWrap = 12, marks = c(",", "."), nDigits = 2, label = NULL,
  percentage = FALSE, popup = NULL, size = c(3, 20), tiles = NULL)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbles by latitud and longitud and numeric variable
}
\section{ctypes}{
 Gln-Glt-Num
}

\examples{
lflt_bubbles_GlnGltNum(sampleData("Gln-Glt-Num", nrow = 10))
}
