% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{lflt_bubbles_GlnGltNum}
\alias{lflt_bubbles_GlnGltNum}
\title{Leaflet bubbles by latitud and longitud and numeric variable}
\usage{
lflt_bubbles_GlnGltNum(data, caption = NULL, agg = "sum",
  border = list(weight = 1.3, color = "black", opacity = 1, stroke =
  TRUE), fill = list(color = NULL, opacity = 0.5, scale = "discrete",
  nullColor = "#dddddd"), format = c("", ""), highlightValue = NULL,
  highlightValueColor = NULL, labelWrap = 12, legend = list(bins = 6,
  position = "bottomleft", title = NULL), marks = c(",", "."),
  nDigits = 2, label = NULL, percentage = FALSE, popup = NULL,
  size = c(3, 20), tiles = NULL)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbles by latitud and longitud and numeric variable
}
\section{ctypes}{
 Gln-Glt-Num
}

\examples{
lflt_bubbles_GlnGltNum(sampleData("Gln-Glt-Num", nrow = 10))
}
