% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles_GlnGlt.R
\name{lflt_bubbles_GlnGltNum}
\alias{lflt_bubbles_GlnGltNum}
\title{#' Bubbles chart Gln Glt Num}
\usage{
lflt_bubbles_GlnGltNum(data = NULL, ...)
}
\description{
\code{lflt_bubbles_GlnGlt()} Create a Leaflet bubbles map based on a particular data type.
In this case, you can load data with only three columns, where the firts column contains \strong{longitudes points},
the second \strong{latitudes points} and the third must be a \strong{numeric class column},
or be sure that three firts columns they meet this condition.
}
\section{Ftype}{

Gln-Glt-Num
}

\examples{
data <- sample_data("Gln-Glt-Num", n = 30)
lflt_bubbles_GlnGltNum(data)
}
\seealso{
Other Gln-Glt-Num plots: 
\code{\link{lflt_heatmap_GlnGltNum}()}
}
\concept{Gln-Glt-Num plots}
