% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{lflt_bubbles_Gnm}
\alias{lflt_bubbles_Gnm}
\title{Leaflet bubbles geo name}
\usage{
lflt_bubbles_Gnm(data = NULL, mapName = "world_countries",
  opts = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbles
}
\section{ctypes}{
 Gnm
}

\examples{
lflt_bubbles_Gnm(sampleData("Gnm", nrow = 10))
}
