% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles_GnmCatNum.R
\name{lflt_bubbles_GnmCatNum}
\alias{lflt_bubbles_GnmCatNum}
\title{Leaflet bubbles by categorical variable}
\usage{
lflt_bubbles_GnmCatNum(data = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbles by categorical variable
}
\section{ctypes}{
 Gnm-Cat-Num
}

\examples{
lflt_bubbles_GnmCatNum(sample_data("Gnm-Cat-Num", nrow = 10))
}
