% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles_GnmNum.R
\name{lflt_bubbles_GnmNum}
\alias{lflt_bubbles_GnmNum}
\title{Leaflet bubbless by numerical variable}
\usage{
lflt_bubbles_GnmNum(data, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbless by numerical variable
}
\section{ctypes}{
 Gnm-Num
}

\examples{
lflt_bubbles_GnmNum(sampleData("Gnm-Num", nrow = 10))
}
