% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{lflt_bubbles_GnmNum}
\alias{lflt_bubbles_GnmNum}
\title{Leaflet bubbles geo name num}
\usage{
lflt_bubbles_GnmNum(data = NULL, mapName = "world_countries",
  opts = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet bubbles
}
\section{ctypes}{
 Gnm-Num
}

\examples{
lflt_bubbles_GnmNum(sampleData("Gnm-Num", nrow = 10))
}
