% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{lflt_bubbles_size_GcdNum}
\alias{lflt_bubbles_size_GcdNum}
\title{lflt_bubbles_size_GcdNum
Multilines}
\usage{
lflt_bubbles_size_GcdNum(data, scope = "world", agg = NULL, bounds = NULL)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
lflt_bubbles_size_GcdNum
Multilines
}
\section{ftypes}{
 Ye-Nu*
}

\examples{
lflt_bubbles_GcdNum(sampleData("Gcd-Num",nrow = 10))
}
