% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth_GcdNum.R
\name{lflt_choropleth_Gcd}
\alias{lflt_choropleth_Gcd}
\title{Leaflet choropleths by numerical variable}
\usage{
lflt_choropleth_Gcd(data = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet choropleths by numerical variable
}
\section{ctypes}{
 Gcd
}

\examples{
lflt_choropleth_Gcd(sampleData("Gcd", nrow = 10))
}
