% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleths.R
\name{lflt_choropleth_Gcd}
\alias{lflt_choropleth_Gcd}
\title{Leaflet choropleths by geographical code}
\usage{
lflt_choropleth_Gcd(data = NULL, caption = NULL, count = TRUE,
  border = list(weight = 1, color = "black", opacity = 1),
  fill = list(color = NULL, opacity = 0.5, scale = "no", nullColor =
  "#cccccc"), format = c("", ""), highlightValue = NULL,
  highlightValueColor = NULL, labelWrap = 12, legend = list(bins = 6,
  position = "bottomleft", title = NULL), mapName = "world_countries",
  marks = c(",", "."), nDigits = 2, label = NULL,
  percentage = FALSE, popup = NULL, tiles = NULL)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet choropleths by geographical code
}
\section{ctypes}{
 Gcd
}

\examples{
lflt_choropleth_Gcd(sampleData("Gcd", nrow = 10))
}
