% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleths.R
\name{lflt_choropleth_Gcd}
\alias{lflt_choropleth_Gcd}
\title{Leaflet choropleths by geographical code}
\usage{
lflt_choropleth_Gcd(data, fillOpacity = 0.5, format = c("", ""),
  legendColor = "discrete", legendPosition = "bottomleft",
  legendTitle = NULL, label = NULL, marks = c(",", "."), nBins = 4,
  nDigits = 2, palette = c("#009EE3", "#E5007D", "#95C11E"),
  percentage = FALSE, scope = "world_countries",
  tiles = "CartoDB.Positron")
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet choropleths by geographical code
}
\section{ctypes}{
 Gcd
}

\examples{
lflt_choropleth_Gcd(sampleData("Gcd", nrow = 10))
}
