% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleths.R
\name{lflt_choropleth_GcdCat}
\alias{lflt_choropleth_GcdCat}
\title{Leaflet choropleths by geographical variable}
\usage{
lflt_choropleth_GcdCat(data = NULL, caption = NULL, count = TRUE,
  border = list(), fill = list(), format = c("", ""),
  labelWrap = 12, legend = list(), mapName = "world_countries",
  marks = c(",", "."), nDigits = 2, label = NULL,
  percentage = FALSE, popup = NULL, tiles = NULL)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet choropleths by geographical variable
}
\section{ctypes}{
 Gcd-Cat
}

\examples{
lflt_choropleth_GcdCat(sampleData("Gcd-Cat", nrow = 10))
}
