% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth_GcdCatNum.R
\name{lflt_choropleth_GcdCat}
\alias{lflt_choropleth_GcdCat}
\title{Leaflet choropleth by categorical variable}
\usage{
lflt_choropleth_GcdCat(data = NULL, ...)
}
\description{
\code{lflt_choropleth_GcdCat()} Create a Leaflet choropleth map based on a particular data type.
In this case, you can load data with only two columns, where the firts it's a
\strong{geocode column}, and second is a \strong{categorical column},
or make sure that the firts two columns of you data meet this condition
}
\section{Ftype}{

Gcd-Cat
}

\examples{
data <- sample_data("Gcd-Cat", n = 30)
lflt_choropleth_GcdCat(data)

# Activate data labels
lflt_choropleth_GcdCat(data,
                       dataLabels_show = TRUE)

# data with more of one column
data <- sample_data("Gcd-Cat-Dat-Yea-Cat", n = 30)
lflt_choropleth_GcdCat(data)

# Change variable to color and pallete type
lflt_choropleth_GcdCat(data,
                       color_by = names(data)[2],
                       palette_type = "sequential")

# Change tooltip info and add additional information contained in your data
names_data <- names(data)
info_tool <- paste0("<b>",names_data[1],":</b> {", names_data[1],"}<br/><b>", names_data[4],":</b> {", names_data[4],"}<br/>")
data \%>\%
 lflt_choropleth_GcdCat(tooltip = info_tool)
}
