% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth_GcdCatNum.R
\name{lflt_choropleth_GcdCat}
\alias{lflt_choropleth_GcdCat}
\title{Leaflet choropleths by categorical variable}
\usage{
lflt_choropleth_GcdCat(data = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet choropleths by categorical variable
}
\section{ctypes}{
 Gcd-Cat
}

\examples{
lflt_choropleth_GcdCat(sample_data("Gcd-Cat", nrow = 10))
}
