% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleths.R
\name{lflt_choropleth_GcdCat}
\alias{lflt_choropleth_GcdCat}
\title{Leaflet choropleths by geographical variable}
\usage{
lflt_choropleth_GcdCat(data = NULL, caption = NULL, count = TRUE,
  border = list(weight = 1, color = "black", opacity = 1),
  fill = list(color = NULL, opacity = 0.5, scale = "discrete", nullColor
  = "#dddddd"), format = c("", ""), highlightValue = NULL,
  highlightValueColor = NULL, labelWrap = 12, legend = list(bins = 6,
  position = "bottomleft", title = NULL), mapName = "world_countries",
  marks = c(",", "."), nDigits = 2, label = NULL,
  percentage = FALSE, popup = NULL, tiles = NULL)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet choropleths by geographical variable
}
\section{ctypes}{
 Gcd-Cat
}

\examples{
lflt_choropleth_GcdCat(sampleData("Gcd-Cat", nrow = 10))
}
