% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleths.R
\name{lflt_choropleth_GcdNum}
\alias{lflt_choropleth_GcdNum}
\title{Leaflet choropleths by numerical variable}
\usage{
lflt_choropleth_GcdNum(data, palette = c("#009EE3", "#9B71AF", "#FFFF99"),
  fillOpacity = 0.5, label = NULL, popup = NULL, agg = "sum",
  scope = "world_countries", tiles = "CartoDB.Positron")
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet choropleths by numerical variable
}
\section{ctypes}{
 Gcd-Num
}

\examples{
lflt_choropleth_GcdNum(sampleData("Gcd-Num", nrow = 10))
}
