% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth_GcdNum.R
\name{lflt_choropleth_GcdNum}
\alias{lflt_choropleth_GcdNum}
\title{Leaflet choropleths by numerical variable}
\usage{
lflt_choropleth_GcdNum(data = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet choropleths by numerical variable
}
\section{ctypes}{
 Gcd-Num
}

\examples{
lflt_choropleth_GcdNum(sampleData("Gcd-Num", nrow = 10))
}
