% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\name{lflt_choropleth_Gnm}
\alias{lflt_choropleth_Gnm}
\title{Leaflet choropleths by geo name}
\usage{
lflt_choropleth_Gnm(data = NULL, mapName = "world_countries",
  opts = NULL)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet choropleths by geo name
}
\section{ctypes}{
 Gnm
}

\examples{
lflt_choropleth_Gnm(sampleData("Gnm", nrow = 10))
}
