% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth_GnmNum.R
\name{lflt_choropleth_Gnm}
\alias{lflt_choropleth_Gnm}
\title{Leaflet choropleths by numerical variable}
\usage{
lflt_choropleth_Gnm(data = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet choropleths by numerical variable
}
\section{ctypes}{
 Gnm
}

\examples{
lflt_choropleth_Gnm(sampleData("Gnm", nrow = 10))
}
