% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleths.R
\name{lflt_choropleth_Gnm}
\alias{lflt_choropleth_Gnm}
\title{Leaflet choropleths by geographical name}
\usage{
lflt_choropleth_Gnm(data, palette = c("#009EE3", "#9B71AF", "#FFFF99"),
  fillOpacity = 0.5, label = NULL, popup = NULL,
  scope = "world_countries", tiles = "CartoDB.Positron")
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet choropleths by geographical name
}
\section{ctypes}{
 Gnm
}

\examples{
lflt_choropleth_Gnm(sampleData("Gnm", nrow = 10))
}
