% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth_GnmCatNum.R
\name{lflt_choropleth_GnmCat}
\alias{lflt_choropleth_GnmCat}
\title{Leaflet choropleths by categorical variable}
\usage{
lflt_choropleth_GnmCat(data = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet choropleths by categorical variable
}
\section{ctypes}{
 Gnm-Cat
}

\examples{
lflt_choropleth_GnmCat(sample_data("Gnm-Cat", nrow = 10))
}
