% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleths.R
\name{lflt_choropleth_GnmCat}
\alias{lflt_choropleth_GnmCat}
\title{Leaflet choropleths by categorical variable}
\usage{
lflt_choropleth_GnmCat(data, palette = c("#009EE3", "#9B71AF",
  "#FFFF99"), fillOpacity = 0.5, label = NULL, popup = NULL,
  scope = "world_countries", tiles = "CartoDB.Positron")
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet choropleths by categorical variable
}
\section{ctypes}{
 Gnm-Cat
}

\examples{
lflt_choropleth_GnmCat(sampleData("Gnm-Cat", nrow = 10))
}
