% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth_GnmCatNum.R
\name{lflt_choropleth_GnmCatNum}
\alias{lflt_choropleth_GnmCatNum}
\title{Leaflet choropleths by categorical variable}
\usage{
lflt_choropleth_GnmCatNum(data = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet choropleths by categorical variable
}
\section{ctypes}{
 Gnm-Cat-Num
}

\examples{
lflt_choropleth_GnmCatNum(sample_data("Gnm-Cat-Num", nrow = 10))
}
