% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleths.R
\name{lflt_choropleth_GnmNum}
\alias{lflt_choropleth_GnmNum}
\title{Leaflet choropleths by numerical variable}
\usage{
lflt_choropleth_GnmNum(data, palette = c("#009EE3", "#9B71AF",
  "#FFFF99"), fillOpacity = 0.5, label = NULL, popup = NULL,
  agg = "sum", scope = "world_countries", tiles = "CartoDB.Positron")
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet choropleths by numerical variable
}
\section{ctypes}{
 Gnm-Num
}

\examples{
lflt_choropleth_GnmNum(sampleData("Gnm-Num", nrow = 10))
}
