% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleths.R
\name{lflt_choropleth_GnmNum}
\alias{lflt_choropleth_GnmNum}
\title{Leaflet choropleths by numerical variable}
\usage{
lflt_choropleth_GnmNum(data, agg = "sum", dropNa = FALSE,
  fillOpacity = 0.5, format = c("", ""), legendColor = "discrete",
  legendPosition = "bottomleft", legendTitle = NULL, label = NULL,
  marks = c(",", "."), nBins = 4, nDigits = 2,
  palette = c("#009EE3", "#E5007D", "#95C11E"), percentage = FALSE,
  scope = "world_countries", tiles = "CartoDB.Positron")
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet choropleths by numerical variable
}
\section{ctypes}{
 Gnm-Num
}

\examples{
lflt_choropleth_GnmNum(sampleData("Gnm-Num", nrow = 10))
}
