% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{lflt_heatmap_GlnGlt}
\alias{lflt_heatmap_GlnGlt}
\title{Leaflet heatmap by longitud and latitud}
\usage{
lflt_heatmap_GlnGlt(data, radius = NULL, blur = NULL, gradient = NULL,
  tiles = "CartoDB.Positron")
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Leaflet heatmap by longitud and latitud
}
\details{
lflt_heatmap_GlnGlt
Heatmap
}
\section{ctypes}{
 Gln-Glt
}

\examples{
lflt_heatmap_GlnGlt(sampleData("Gln-Glt", nrow = 10))
}
